Added a new RH/T sensor shield
This shield is designed to be used with different sensors depending on the
requirements. It also comes with a 8P8C (RJ45) connector, connected to the same
buses than the sensors on the board, in order to be able to put sensors further
away from the bugOne.

A repeater board is provided to connect the sensors with the RJ45 connectors.