Adding "Sleep announce" message
This message allows proper "node presence" management in the BugNet. Before
going in sleeping mode, the node sends a message to warn everyone that it will
not be available. This is especially useful for nodes able to receive commands.
Commands can be buffered until the node comes back. Receive only device should
send a "Hello" when waking up